rename server to cjs and update script
