+++ myapp/package.json
@@ -8,7 +8,7 @@
     "build": "tsc",
     "preview": "vite preview",
     "test": "node ./test.js",
-    "server": "node ./server.js"
+    "server": "node ./server.cjs"
   },
   "dependencies": {
     "bootstrap": "^5.3.0",
